+++ src/System.ObjectModel/tests/System.ObjectModel.Tests.csproj
@@ -23,6 +23,7 @@
       <Link>Common\System\CollectionsIDictionaryTest.cs</Link>
     </Compile>
     <Compile Include="ObservableCollection\ObservableCollection_ConstructorAndPropertyTests.cs" />
+    <Compile Include="ObservableCollection\ObservableCollection_MethodsTest.cs" />
     <Compile Include="ObservableCollection\ObservableCollection_ReentrancyTests.cs" />
     <Compile Include="ObservableCollection\ObservableCollection_Serialization.cs" />
     <Compile Include="System\Collections\ObjectModel\KeyedCollectionTests.cs" />
@@ -43,7 +44,7 @@
     </Compile>
   </ItemGroup>
   <ItemGroup Condition="'$(TargetGroup)'=='netcoreapp'">
-    <Compile Include="ObservableCollection\ObservableCollection_MethodsTest.netcoreapp.cs" />
+    <Compile Include="ObservableCollection\ObservableCollection_MethodTests.netcoreapp.cs" />    
     <Compile Include="ObservableCollection\ObservableCollection_SkipCollectionChangedTests.netcoreapp.cs" />
     <Compile Include="KeyedCollection\TestMethods.netcoreapp.cs" />
     <Compile Include="System\Collections\ObjectModel\KeyedCollectionTests.netcoreapp.cs" />
